requirements.txt
azure-cognitiveservices-speech==1.37.0
azure-ai-textanalytics==5.3.0
python-dotenv==1.0.1
pymongo==4.7.2
langchain==0.2.2
langchain-community 
transformers==4.41.2
tensorflow==2.16.1
openai==0.28.0
gradio==4.36.1
soundfile==0.12.1
pytest==8.2.2